requirements.txt
requests
spotdl
pyTelegramBotAPI
python-dotenv